requirements.txt
requests
tweepy
feedparser
python-dotenv
supabase
scikit-learn
sentence-transformers
torch
openai
